+++ src/application/Window.h
@@ -53,7 +53,7 @@ public:
 
 	bool Init(const std::string  &title, uint32_t width, uint32_t height, bool fullscreen = false, bool windowed = true, bool vsync = false, uint32_t msaa = 0, uint32_t r = 8, uint32_t g = 8, uint32_t b = 8, uint32_t alpha = 8, uint32_t depth = 24, uint32_t stencil = 8);
 	bool GetKey(uint32_t key);
-	glm::ivec2 GetMousePos();
+	glm::vec2 GetMousePos();
 	glm::ivec2 GetWindowSize();
 	void SetWindowSize(uint32_t width, uint32_t height);
 	void SetWindowMode(bool fullscreen, bool windowed);
